feat(notebooks): add knn modeling
Created a vanilla baseline knn model, with no hyperparamters tuned,
we did set the n_jobs to -1 just so that we could use all processors
when running predictions.

Performing an exhaustive search over the entire parameter space would
require too much time, using RandomizedSearchCV, we did a random search
of the parameter space to find an optimal parameter set.

Final conclusion is that knn is not optimal for this problem.