feat(notebooks): add naive bayes modeling
Basic model without accounting for class imbalance has resulted in a
model with poor performance in both training and testing. Accuracy is
below 50% in both training and testing.

Our new model adjusted for class imbalance sees a boost in accuracy in
both training and testing, although precision and recall are still low
on testing data. Next steps will be to use feature selection to reduce
our dimensions and try to find a good model fit.

Final Model with 6 features performs decent in both testing and training,
will have to see performance on the final testing dataset.