feat(notebooks): add benchmark modeling
Tested vanilla models with no hyper parameter tuning on both our
imbalanced and balanced training data. Found that the classifier that
performed the best was a Gradient Boost Random Forest model on
imbalanced data. It had an f1_score of .77 on testing data, and an AUC
score of .88 on testing data.

Moving forward we'll use this model and seek to tune it to better
perform.